api/requirements.txt
firebase-admin==6.5.0
Flask==3.0.3
flask-cors==4.0.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
yfinance==0.2.40
python-gemini-api==2.4.12
jwt==1.3.1
pandas==2.2.2